refactor - wishlist data array updated
