+++ src/data/wishlistImg.data.js
@@ -7,6 +7,28 @@ import forumlS from '../assets/3-s.jpg'
 import keithcM from '../assets/4-m.jpg'
 import keithcS from '../assets/4-s.jpg'
 
-export {
-    am97M, am97S, jpsM, jpsS, forumlM, forumlS, keithcM, keithcS
-}
+export const wishlistImgArr = [
+    {
+        id: 1,
+        alt: 'air max 97',
+        srcSet: `${am97M} 200w, ${am97S} 150w`,
+    },
+    {
+        id: 2,
+        alt: 'japan s',
+        srcSet: `${jpsM} 200w, ${jpsS} 150w`,
+    },
+    {
+        id: 3,
+        alt: 'forum luxe',
+        srcSet: `${forumlM} 200w, ${forumlS} 150w`,
+    },
+    {
+        id: 4,
+        alt: 'keith haring club c',
+        srcSet: `${keithcM} 200w, ${keithcS} 150w`,
+    }
+
+]
+
+export const sizes = '(max-width: 768px) 150px, 200px'
